% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluatr.crossval}
\alias{evaluatr.crossval}
\title{Perform cross-validation}
\usage{
evaluatr.crossval(analysis)
}
\arguments{
\item{analysis}{Analysis object, initialized by TODO.init. You must call TODO.impact before calling TODO.sensitivity}
}
\value{
Cross-validation results, `results`, as described below

`results$full` TODO
`results$time` TODO
`results$time_no_offset` TODO
`results$pca` TODO
`results$ann_pred_quantiles_stack` TODO
`results$cumsum_prevented_stack` TODO
`results$log_rr_quantiles_stack` TODO
`results$log_rr_samples.prec.post_stack` TODO
`results$point.weights` TODO
`results$pred_quantiles_stack` TODO
`results$quantiles_stack` TODO
`results$rr_mean_combo` TODO
`results$rr_mean_stack` TODO
`results$rr_mean_stack_intervals` TODO
`results$rr_roll_stack` TODO
`results$stacking_weights` TODO
`results$stacking_weights.all` TODO
`results$stacking_weights.all.m` TODO
}
\description{
Perform cross-validation
}
