% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluatr.impact}
\alias{evaluatr.impact}
\title{Perform impact analysis}
\usage{
evaluatr.impact(analysis, variants = names(analysis$.private$variants))
}
\arguments{
\item{analysis}{Analysis object, initialized by evaluatr.init.}

\item{variants}{List of one or more of the following analysis types to perform: 'time', 'time_no_offset', 'pca', and 'full'. Defaults to all 4.}
}
\value{
Analysis results, `results`, as described below

`results$full` Results from synthetic controls model

`results$time` Results from time trend model (with offset term)

`results$time_no_offset` Results from time trend model (no offset term)

`results$pca` Results from STL+PCA model

`results$its` Results from classic interrupted time series model

`results$rr_mean_combo` Rate ratios during the evaluation period, as calculated in each of the models.
}
\description{
Perform impact analysis
}
