% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluatr.plan}
\alias{evaluatr.plan}
\title{Construct an evaluation plan for the analysis. Use this with future::plan to control how InterventionEvaluatR uses multi-core and cluster computation resources}
\usage{
evaluatr.plan(analysis)
}
\arguments{
\item{analysis}{Analysis object, initialized by evaluatr.init.}
}
\value{
Future evaluation plan, suitable for passing to future::plan
}
\description{
Construct an evaluation plan for the analysis. Use this with future::plan to control how InterventionEvaluatR uses multi-core and cluster computation resources
}
