% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{evaluatr.prune}
\alias{evaluatr.prune}
\title{Prune unneeded data from the analysis object. Used internally before saving to reduce size of saved data.}
\usage{
evaluatr.prune(analysis)
}
\description{
Prune unneeded data from the analysis object. Used internally before saving to reduce size of saved data.
}
\keyword{internal}
