% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluatr.sensitivity}
\alias{evaluatr.sensitivity}
\title{Perform sensitivity analysis}
\usage{
evaluatr.sensitivity(analysis)
}
\arguments{
\item{analysis}{Analysis object, initialized by TODO.init. You must call TODO.impact before calling TODO.sensitivity}
}
\value{
Sensitivity analysis results, `results`, as described below

`results$rr_table` TODO
`results$rr_table_intervals` TODO
`results$sensitivity_pred_quantiles` TODO
`results$sensitivity_table` TODO
`results$sensitivity_table_intervals` TODO
}
\description{
Perform sensitivity analysis
}
