% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluatr.univariate}
\alias{evaluatr.univariate}
\title{Perform analysis controling for 1 variable at a time}
\usage{
evaluatr.univariate(analysis)
}
\arguments{
\item{analysis}{Analysis object, initialized by TODO.init.}
}
\value{
Univariate analysis results, `results`, as described below

`results$rr` Rate ratio (RR) estimate (median) from single control model

`results$rr.ucl` RR Upper 95% CrI from single control model

`results$rr.lcl` RR Lower 95% CrI from single control model
 
`results$aic.wgt` Akaike weight of the univariate model

`results$covar` Name of the control variable used for adjustment
}
\description{
Perform analysis controling for 1 variable at a time
}
