% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy-package.R
\docType{package}
\name{InterventionEvaluatR-package}
\alias{InterventionEvaluatR-package}
\title{TODO}
\description{
One sentence package summary TODO
}
\details{
More detail TODO
...
}
\author{
Daniel Weinberger \email{daniel.weinberger@yale.edu}
}
\keyword{TODO}
