% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy-data.R
\docType{data}
\name{pnas_chile}
\alias{pnas_chile}
\title{TODO}
\format{Monthly time series of cause-specific hospitalizations from Chile, as described in Bruhn et al., PNAS. 
The columns refer to sets of ICD10 codes. 
'age_group' categories: 
  2:12-23m 
  3:24-59m 
  4:5-17 years
  5: 18-39 years
  6:40-64 years
  7:65-79 years
  8: 80+ years
  92: <24 months}
\description{
TODO
}
