% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{syncon.plots}
\alias{syncon.plots}
\title{Generate plots for TODO}
\usage{
syncon.plots(analysis)
}
\arguments{
\item{analysis}{Analysis object, initialized by TODO.init.}
}
\value{
Plots, `plots`, as described below

`plots$summary` TODO
`plots$groups` TODO
}
\description{
Generate plots for TODO
}
