% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{syncon.save}
\alias{syncon.save}
\title{Save data for TODO}
\usage{
syncon.save(analysis, output_file)
}
\arguments{
\item{analysis}{Analysis object, initialized by TODO.init.}

\item{output_file}{An Rds file in which analysis results will be saved.}
}
\description{
Save data for TODO
}
