% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluatr.impact}
\alias{evaluatr.impact}
\title{Perform impact analysis}
\usage{
evaluatr.impact(analysis)
}
\arguments{
\item{analysis}{Analysis object, initialized by TODO.init.}
}
\value{
Analysis results, `results`, as described below

`results$full` TODO
`results$time` TODO
`results$time_no_offset` TODO
`results$pca` TODO
`results$its` TODO
`results$best` TODO
`results$point.weights` TODO
`results$rr_mean_combo` TODO
}
\description{
Perform impact analysis
}
