% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{evaluatr.save}
\alias{evaluatr.save}
\title{Save data for TODO}
\usage{
evaluatr.save(analysis, output_file, prune = TRUE)
}
\arguments{
\item{analysis}{Analysis object, initialized by TODO.init.}

\item{output_file}{An Rds file in which analysis results will be saved.}

\item{prune}{If TRUE (default) then diagnostic data is not saved. If FALSE, diagnostic data is saved (making the output file much larger).}
}
\description{
Save data for TODO
}
