% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_aux.R
\name{evaluatr.univariate.plot}
\alias{evaluatr.univariate.plot}
\title{Generate plot of univariate results}
\usage{
evaluatr.univariate.plot(ds)
}
\arguments{
\item{ds}{Object imported from evaluatr.univariate}
}
\value{
Univariate analysis plot, `results`, as described below
}
\description{
Generate plot of univariate results
}
